+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+- Post-quantum cryptography is now enabled by default
+- [Feature] Add functions to get default VPN settings and free allocated resources.
+    - See `vpn_get_default_settings` and `vpn_free_default_settings`.
+
 ## 0.99.96
 
 - [Fix] [Windows] Fixed certificate validation failing with `WCRYPT_E_POLICY_STATUS` error.